% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedates.r
\name{makedates}
\alias{makedates}
\title{Make a standard table of dates from the R datelist you feed it}
\usage{
makedates(dates.in)
}
\arguments{
\item{infile}{A vector of dates, or a character vector in the format "yyyy-mm-dd"}
}
\value{
A data frame of all the date attributes
}
\description{
Sometimes it is nice to just extract the attributes of an R date-list rather than messing with the date format
This takes a vector of dates and outputs a data.frame containing the following
}
\details{
ColumnName (class)    : Description (strptime code as appropriate)

 - Date         (Date)       : Date in yyyy-mm-dd format.
 - Year         (numeric)    : Year (as.numeric(%Y))
 - Month        (numeric)    : Month (as.numeric(%m))
 - Day          (numeric)    : The day in the month (as.numeric(%d))
 - Dekad        (numeric)    : The 10 day period (3 per month), from 1-36 (custom)
 - Pentad       (numeric)    : The 5 day period (6 per month), from 1-72 (custom)
 - DOY          (numeric)    : The Julian Day of Year e.g. 365 in a non-leap year, 366 in a leap year (as.numeric(%j))
 - DOY366       (numeric)    : The 5 day period (6 per month), from 1-72 (custom)
 - MonthDay     (numeric)    : A number of the month-day (mmdd). Essentially a numeric version of DOY366  (mmdd)
 - YearMonth    (numeric)    : A number of the year-month (yyyymm). Useful for monthly splits/group statistics (yyyymm)
 - YearDekad    (numeric)    : A number of the year-dekad (yyyyDD). Useful for dekadal splits/group statistics  (yyyyDD)
 - YearPentad   (numeric)    : A number of the year-pentad (yyyypp). Useful for pentadal splits/group statistics (yyyyPP)
 - Season       (character)  : The boreal season (DJF,MAM,JJA,SON) e.g. DJF is December,January,February
 - SomaliSeason (character)  : Rough Somali seasons, e.g. Dehr (Gu, ,Dehr,jilaal)
}
