% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettemaker.R
\name{palettemaker}
\alias{palettemaker}
\title{palettemaker}
\usage{
palettemaker(
  package,
  pal,
  n,
  reverse = FALSE,
  removestart = FALSE,
  removeend = FALSE,
  rgb = FALSE
)
}
\arguments{
\item{package}{currently either "brewer" or "viridis"}

\item{pal}{The sub-palette you want e.g. "Blues"}

\item{n}{How many output colours you want}

\item{reverse}{Set to true if you want the colors to go the other way}

\item{removestart}{numeric: Setting to 1 removes the first color, 2 the two first colors etc.}

\item{removeend}{numeric: Setting to 1 removes the final color, 2 the two final colors etc.}

\item{rgb}{Set to true to output rgb values rather than hex}
}
\value{
It outputs a list, with the first value a description of the options, and the second a list value with the colors in either hex or rgb.
}
\description{
Create a directory if it doesn't already exist
}
\details{
Someone wanted the rgbs for a custom 19 value colorbrewer color-scale. This is the result of that request.
}
\examples{
#-------------------------------------------------------------------------------------
# Example 1: Basic hex
#-------------------------------------------------------------------------------------
   Example1 <- palettemaker(package = "viridis",  pal = "viridis",
                              n  = 10,  rgb = FALSE, reverse = TRUE,
                              removeend = 0, removestart = 2)

   #The first list value is the option you entered
   #The second list value are the colours
    print(Example1[[1]])
    print(Example1[[2]])

   #Make a plot
    plot(1:10, 1:10,cex=2,pch=16,col= Example1[[2]],main=Example1[[1]])

#-------------------------------------------------------------------------------------
# Example 2: Create a palette, save the colours to the palette code and plot
#-------------------------------------------------------------------------------------
   Example2 <- palettemaker("brewer","Blues",17,rgb=FALSE,reverse=FALSE,removeend=0,removestart=0)
   eval(parse(text=paste(paste(Example2[[1]], "<- Example2[[2]]"))))
   plot(1:17, 1:17,cex=2,pch=16,col= brewer_Blues_17_hex)

#-------------------------------------------------------------------------------------
# Example 3: Make an rgb output and save output text file to the Output subfolder
#            Note your final palette will be of size n,
#            even if you remove colours from the original palette
#-------------------------------------------------------------------------------------

   rgbpal <- palettemaker("brewer","Spectral",22,rgb=TRUE,reverse=TRUE,removeend=1,removestart=3)
   filename <- paste(rgbpal[[1]],".txt",sep="")
   write.table(rgbpal[[2]],filename,col.names = FALSE,row.names = FALSE,quote=FALSE,sep = "\t")
}
