% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterstack_mean_fast.R
\name{rasterstack_mean_fast}
\alias{rasterstack_mean_fast}
\title{Stack a raster quickly and take the mean}
\usage{
rasterstack_mean_fast(x)
}
\arguments{
\item{infile}{A raster stack}
}
\value{
A single raster which is the mean of the stack that was input
}
\description{
When running raster stacks in parallel this is the fastest way I have found
All credit to Samuel Bosch
I found this code at
http://www.samuelbosch.com/2018/02/speeding-up-mean-and-standard-deviation.html
and adapted it to just take the mean
}
