% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JumpDate.r
\name{JumpDate}
\alias{JumpDate}
\title{JumpDate}
\usage{
JumpDate(
  DateIn = NA,
  TimeBack = NA,
  TimeUnit = NA,
  Direction = "back",
  verbose = FALSE
)
}
\arguments{
\item{DateIn}{Date: An R-Date, or something that can be converted automatically by as.Date}

\item{TimeUnit}{string: String. Anything that works in seq.Date e.g. "years" "months" "days"}

\item{Direction}{string: "back" for back, everything else goes forward}

\item{verbose}{TRUE/FALSE. Do you want it to print out what you are doing? Useful for debugging}

\item{TimeJump}{numeric: Number of units back/forward e.g. 3 days, TimeJump=3}
}
\value{
Your new date
}
\description{
Jump forward/backwards X number of days/weeks/months in base R.
Sometimes you don't want to bother with packages
}
\details{
I want to be able to easily move forward and back in time without having to use a package
See ?seq.Date for more examples of what you can add in TimeUnit.

See here for other ways to achieve this and the background to this function
https://stackoverflow.com/questions/5225823/how-to-subtract-months-from-a-date-in-r
}
\examples{
# Move 1 month back from 2000-02-29
JumpDate("2000-02-29",2,"months","back")

# Move 14 weeks forward from Jan-1-2002 and print out all output
JumpDate(DateIn="2002-01-01",TimeBack=14,TimeUnit="weeks",Direction="forward",verbose=TRUE)
}
